Catalogs restructuralization + css modules using
