+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/react-redux": "^7.1.21",
+    "classnames": "^2.3.1",
     "history": "^5.2.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
