+++ src/app/App.css
@@ -4,4 +4,11 @@
   color: white;
   border: 2px solid gray;
   border-radius: 20px;
+  width: 100%;
+}
+
+.container {
+  width: 100%;
+  max-width: 650px;
+  margin: 0 auto;
 }
