+++ src/fonts/FontList.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { select } from './fontSlice';
+import styles from './Fonts.module.css';
+
+export function FontList() {
+  const fontList = useSelector(state => state.fonts)
+  const dispatch = useDispatch()
+
+  return (
+    <ul className={styles.list}> 
+      {fontList.map((item) => 
+        <li key={item.id} 
+        className={[styles.block,  item.isActive ? styles.active : ""].join(" ")} 
+        onClick={() => dispatch(select({id: item.id}))}>
+
+          <span 
+          className={styles.color} 
+          style={{backgroundColor: item.color}}>
+            {item.abbr}
+          </span>
+
+          <p>{item.label}</p>
+        </li>
+      )}
+    </ul>
+  )
+}
